line chart: drawable and data conversion
This change introduces a concept of a drawable. Drawable is a region in
a view where an implementation can draw shapes using PaintBrush based on
coordinate converted data.

DataDrawable manages data (and its coordinate system conversion) and
render caching so consumer who is implementing a novel representation of
the data can simply focus on drawing operations.